Explicitly separating indices from ngrams
This does not change the way the app functions it just bares this data out for use.  However, I've noticed a critical mistake I made...I have the n-grams that are in both but not the indices from both, only from the summary.   Its currently printing the summary and highlighting where spans are that are also found in the source.